+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore(active-directory): update sdk v3
 - chore(audit-manager): update sdk v3
 - chore(acm): update sdk v3
+- chore(constructs): directconnect constructs update sdk v3
 
 ## [1.12.0] - 2025-04-02